Feature/s12 p31 b106 282 create comment/backend

See merge request s12-final/S12P31B106!255

+++ backend/src/main/java/com/voiceprint/backend/api/comment/dto/CommentListWithCursorDTO.java
@@ -11,5 +11,5 @@ public class CommentListWithCursorDTO {
     private int code;
     private String message;
     private List<CommentGetResponseDTO> comments;
-    private long nextCursor;
+    private Integer nextCursor;
 }
